backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Automation & Input Control
pyautogui==0.9.54
pynput==1.7.6
pygetwindow==0.0.9

# Image Processing & Recognition
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3

# OCR
paddleocr==2.7.0.3
paddlepaddle==2.5.2

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1

# Packaging
pyinstaller==6.2.0

# CORS
python-cors==1.0.0
